Rename singleget package. Improve cd.chans in cache.go
